Initial commit for the conducting atmosphere test. Converges at 2nd order with the HARM algo (ie. no conduction). The implementation has a few hacks that can possibly be done away with by making some "fundamental" changes to the way KHARMA deals with boundaries.
Currently, a global ParArrayND is defined in conducting_atmosphere.hpp and declared in main.cpp as an 'extern'. This variable is necessary to store radial boundary values of the problem that are read from the .txt files and called while applying the Dirichlet boundary conditions (see dirichlet_bc() in conducting_atmosphere.hpp). In future one could have a KBoundaries class whos
e member functions are the various problem-specific boundary conditions. The problem currently throws up an error at the end of program execution since the object is destroyed prior to the decrement count, but this can be ignored.