Fix a bug resulting in bad ImEx driver behavior
In recent dev versions of KHARMA, divB was sometimes climbing next to
ghost zones, in the last physical corner of the grid.  This does not affect
'stable'.
This was because PtoU was not being applied over ghost zones after the
boundary sync, potentially leading to use of old conserved variable data
when calculating zone updates.