EMHD fixes, Flag stuff
This also looks bigger than it is...
1. Fixes ctop of NaN when evolving EMHD
2. Drive-by math optimizations
3. Improve Flag().
Flag() calls must now be matched to EndFlag(),
and produce indented output for readability.
Zone monitors can still be implemented, but must be
(as is likely more convenient) file outputs w/ a new
function OutputNow().

This does *not* fix non-convergence of conducting atmo.
Problem converges with higher_order_terms off,
with or without evolving dP. Does not converge if
higher order terms are enabled, with or without P
->therefore somewhere in q higher-order stuff.