Fix a nasty SMR/AMR artifact
Parthenon's conventions for prolongation operators call over a different
domain based on the values which determine the prolonged value.
In our case despite working from faces, we are preserving divB in cells,
so should be called over the same domain as the default prolongation
operator, not over a domain tied to the particular face we're working on

This was truly awful to chase down, but at least it's preventable going
forward -- documentation of the operator interface will start to solve
these sorts of issues.